1690_efficient_bilateral_cross_moda-Supplementary Material/MM-ID1690-code/requirements.txt
easydict==1.10
faiss_gpu==1.6.4
infomap==2.7.1
numpy==1.21.1
Pillow==9.3.0
Pillow==9.5.0
PyYAML==6.0
PyYAML==6.0
scikit_learn==1.2.0
scipy==1.10.1
tensorboardX==2.5.1
tensorboardX==2.6
torch==1.13.0
torchvision==0.14.0
tqdm==4.61.2
